Sync fingerd with FreeBSD:
WARNS2 cleanup.

When the -p flag is specified, set an environment variable to the name
of the remote host (or rather, the name as mangled by realhostname_sa())
so that the process can use it to behave differently depending on the
origin on the request.  We use this to implement rudimentary visibility
control on our user information.

Remove register keyword.

ANSI function declarations.

Remove (void) casts.