Fix a bug in the installer.
The installer tried to unmount all mount points in order (e.g. /var/tmp
should be unmounted before /var).  The code had a bug which prevent it from
doing it correctly.  It seems that nobody noticed the bug, because the default
file system layout does not suggest subdir mounts.  With the introduction of
PFS and subdir mounts (e.g. a /var/crash PFS), the bug showed up.

Simplify the code a bit and sort the mounted file systems in reverse
lexicographical order.

In-collab-with: mneumann@