Fix a VMWare related bug in the installer
Someone noticed that the installer generated a wrong ifconfig_<if> line
in /etc/rc.conf if used under VMWare and with the lnc(4) driver.  Strange
about that is, that this only happens with that particular driver.  My em(4)
interface under VMWare is detected properly and the rc.conf entry is
correct.

The value of the interface variable is corrupted somewhere between the
invocation of dhclient and the asprintf.  This commit saves the interface
name in string directly after the dhclient invocation which fixes the
problem described above.

This commit is more a work-around than a real fix.  If someone comes up with
a better fix, I'm happy to back this out and test the new one.  Until now, this
makes the network configuration in the installer with VMWare and lnc(4) work.