lpr: don't rely on dirs having st_size > 0
getq determines the initial size of its queue array based on the st_size
of the queue directory.  If the queue directory is on a HAMMER file
system, this size will be reported as 0.  However the following array
growing logic just doubles the array size, thus always staying at 0.

Use a sensible minimum array size of 20 items (~ 512/24).

Reported-and-fix-by:  Patrick Georgi <patrick@georgi-clan.de>