Fix another process stuck-during-exit bug - stuck due to signal.
Fix a bug where an exiting process becomes a zombie, entering the SZOMB
state, but then receives a stop signal before it can be reaped.  If this
occurs the process was incorrectly placed in SSTOP and then later made
SACTIVE again, preventing it from being reaped.

Symptoms were: process shows up in ps in a DEL state but does not show
up in kgdb's 'info thread' list.  Process is the zombie list and is still
a child of init, but it's state is SACTIVE instead of SZOMB and p_xstat
(last signal taken) is SIGSTOP or SIGTSTP.