nrelease: Switch to building pkgsrc packages from source.
The pkgsrc tree that shall be used must be specified via PKGSRC_PATH.

So far the LiveCD build has been successfully tested. The LiveDVD might
still need some love, but I'm committing this now so work on the release
can begin.

Still on the TODO is using binary packages optionally.

Requested-by: dillon