sh: export the default PATH instead of keeping it private
sh specifies a default setting for PATH if none is set in the
environment, but up to now it didn't export this setting.  This leads to
strange effects, like sh being able to execute binaries in /usr/sbin,
but which(1) not being able to find them.  Export PATH to create more
symmetry and harmony.