pxe/tftpboot work - Allow multiple responses to broadcast query
The PXE boot loader may use a broadcast address to query for files via
TFTP, resulting in multiple responses.  Accept the first DATA response
and keep track of the last ERROR response if no DATA response is received.

The first DATA packet locks the server IP for the tftp transfer.

Bring in some structure options (__packed, etc) from FreeBSD plus a fix
to an address increment in the code.