LiveDVD: Fix installer when it is run in the console.
The LiveDVD was recently changed to not auto-start X11 anymore, meaning
the installer can now be run from within X11 (after manually starting
it) or from the console. However, the latter case requires ttyv1 to be
free because the console installer will log stuff on ttyv1. When started
from X11, this output goes to a different window (which is why ttyv1
was configured normally on the LiveDVD).

Reported-by:   matthias
Dragonfly-bug: <http://bugs.dragonflybsd.org/issue1289>