VKERNEL - Fix entry into ddb when running a SMP vkernel.  Fixes stop/cont
The vkernel was trying to send individual LWPs a SIGSTOP, but POSIX handles
SIGSTOP differently from other signals and it was not having the desired
effect.

Instead the vkernel now sends SIGXCPU to each individual LWP and installs
a signal handler to pause the thread until the debugger wishes to wake it up
again.

Reported-by: corecode