Add a build system for amd64 build/test environments in /usr/src/test/amd64
This build system will fetch, compile, and install qemu, do all
necessary buildworlds and buildkernels, including the amd64 cross-builds,
and generate a fully bootable root.img.  The Makefile also includes a
run target to run qemu on the image.

It can take a few hours to build the whole mess due to the buildworlds but
it's a hit-the-start-button-and-forget build which should prove useful.

WARNING: Initial commit only.  qemu 0.10.{0,1} is seriously broken and is not
yet able to fully-boot either a 32 or 64 bit kernel without seg-faulting.