Add cxm(4), a driver for Hauppauge PVR-250/350 video capture cards.
Before the driver can be used, two firmware files have to be extracted
from the Windows driver which cannot be redistributed with DragonFly.
For further information see the manual page.

It was tested on a PVR-250 which was kindly donated by Hasso Tepper
(many thanks for this!).

Ported-from: FreeBSD (ports/multimedia/pvr250), with modifications