tcp/ip4 mapped address: Add tcp6_soport()
For IPv6 v6 only address or inp's address family is not known yet,
e.g. before connect(2) is called on the INET6 socket, this function
acts exactly the same as cpu0_soport() (the soport function before
this commit).  If a INET6 socket is connected to IPv4 mapped address,
then this function simply falls back to tcp_soport().