Add a dummy offset to the arrays generated by genassym to avoid ary[0]
The dummy offset avoids the generation of dummy arrays of size zero.
This whole code path is a hack, but after a lot of messing around
Alex and I determined that it was easier to hack it then to try to
redo the code due to complications introduced by cross-compiled
environments.

Submitted-by: Alex Hornung <ahornung@gmail.com>