ncal: always generate all lines for a month in cal mode
If a month did not span 6 lines, but less, the lines up to line #6 would
not be padded with spaces.  In the year view, if a month on a column to
the left exhibited this state, and a month to the right would span 6
lines, the days in the 6th line of the right month would appear warped and
not aligned under the month they belong to.  Fix this by always generating
6 lines.

Reported-by:  swildner@