undo: always collect history of the file
Commit b9a33d3fb1235626615b96b4b586ce7a7b8386fd added file histories based
on the history of the parent directory, in order to list history fragments
even when the inode changed.  However it failed to collect the history for
the most recent version of the file.  As a result, if all versions of the
parent directory happened to preceed the oldest version of the file, undo
would not collect any history for the file and fail with error 0.

Unconditionally include the history of the file to fix this problem.