etc/rc.d/sendmail: Actually execute sendmail_precmd().
This will cause the sendmail rc script to care about running
newaliases(1) itself if necessary (database doesn't exist or
/etc/mail/aliases is newer than /etc/mail/aliases.db).

Also add a new rc.conf variable 'sendmail_rebuild_aliases' to
turn off this behavior which is on by default.

This means we don't need to run newaliases(1) ourselves in a
couple of places (such as when building the LiveCD), so remove
those.

Dragonfly-bug: <http://bugs.dragonflybsd.org/issue52>