drm: adapt and use drm_close
While importing the last drm code, drm_close was not properly ported and
integrated.  Thus, opens only incremented the reference count, but code
for closes and decrementing the reference count never was executed.

As a result, Xorg drivers, which during probing drm open and close the
device several times, would not be able to perform certain operations
due to a wrong reference count.