ugensa(4): Fix handling of Huawei modems.
Basically move the calling of the reattach code to get the Huawei
modem out of mass storage mode into the _match() routine and
simplify the logic behind when calling it and when not.

Now it is called for any Huawei modem if it has only one interface,
probably only a mass storage interface. Normally this mass-storage
interface is used under windows as it carries the windows-drivers
for the device. As we don't need it, we just tell the box that we
want the modem interface and not mass storage.

Dragonfly-bug: <http://bugs.dragonflybsd.org/issue1324>
Submitted-by:  Alex Hornung <ahornung@gmail.com>
Tested-by:     Brendan Kosowski <bmk@bmk.com.au>
Approved-by:   hasso