USB - Fix polling code
The polling code now calls the intr function instead of trying
to optimize/shortcut the pending-interrupt test.  There are
additional conditions which need action which the test was
not covering.

The polling code now calls the softintr function as well,
as it may be non-operational when the polling code
is being used.