Add a wrapper script to allow "foreign" compilers into our build system.
With this commit, all CCVER values not associated with a compiler in
base will be relegated to the wrapper script which then (in conjunction
with the compilers.conf file) determines which non-base compiler to
use.

So far, defaults for clang (built from SVN) and pkgsrc/wip/gcc44 are
provided.

For more info see the compilers.conf(5) manual page and my mail on
kernel@:

http://leaf.dragonflybsd.org/mailarchive/kernel/2009-06/msg00013.html