CAM - disable cache synchronization by default for USB attachments.
Instead of trying to blacklist particular usb mass storage devices which
do not support the cache synchronization command, just default to not
supporting it for ALL usb mass storage attachments.  Too many usb devices
brick when sent the command.

It matters more now that people are using HAMMER, and HAMMER utilizes the
cache synchronization command to delineate flushes.  We will now by
definition not be able to guarantee recovery if HAMMER is used on a USB
mass storage device.  It will probably work anyway, though.

Reported-by: Many people